refactor: 리턴 문 내 단일 쿠폰 아이템 영역 컴포넌트로 분리
